Move DictifiedV2Route and v2prettyroute to v2route.py
